+++ app/src/main/res/layout/fragment_main_gps.xml
@@ -33,14 +33,9 @@
         android:layout_width="200dp"
         android:layout_height="wrap_content"
         android:layout_above="@+id/gps_text"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="155dp"
-        android:layout_marginTop="26dp"
-        android:layout_marginEnd="156dp"
-        android:layout_marginBottom="52dp"
-        android:onClick="combine"
+        android:layout_centerHorizontal="true"
+        android:background="@drawable/main_list"
+        android:onClick="onClick"
         android:text="서버에 한꺼번에 보내기"
         android:textColor="#FFFFFF"
         android:textSize="10dp" />
