+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Thu Aug 20 01:26:20 KST 2020
-sdk.dir=C\:\\Users\\09jie\\AppData\\Local\\Android\\Sdk
+#Thu Aug 20 02:54:43 KST 2020
+sdk.dir=C\:\\Users\\unja0\\AppData\\Local\\Android\\Sdk
